Add NodeArray to core node class
Restructure of core classes to support two container types, a dictionary
class and an array. The dictionary node is equivalent to the original
Node class so it retains the name. NodeArray contains unnamed children
which are given the index names [0],[1]..[n].

This is a breaking change so this initial commit only fulfils the unit
tests for test_node.py. As many of the public attributes are now
private, new accessors have been added. Also the order of arguments to
the core node class has been swapped from (name,parent) to (parent,name)
which will cause problems in the value classes, and is better addressed
in a separate commit.