Export nodes and values to json dict
Use the node visitor to build a tree of json dictionaries along with
dictionaries of any set values and source connections. The original
intention was to use JSONEncode and JSONDecode, but the tree needs to be
loaded and saved from the root, not from the leaves as in JSONDecode.
Also having to maintain a list of exportable nodes required a 'registry'
that specifies what can be saved and loaded.

Also for type safety exported files require concrete value types, with
standard types in std_value.py for int, float, string and bool. Custom
value types may also be specified in the registry.